+++ lib/phx_live_storybook/attr.ex
@@ -1,10 +1,10 @@
 defmodule PhxLiveStorybook.Attr do
   @moduledoc """
-  An attr is one of your component attributes. Itss structure mimics
+  An attr is one of your component attributes. Its structure mimics
   forthcoming LiveView 0.18.0 declarative assigns.
 
-  Attributes declaration will populate the Documentation tab of your
-  storybook, for each of your components.
+  Attributes declaration will populate the Playground tab of your storybook,
+  for each of your components.
 
   Supported keys:
   - `id`: the attribute id (required). Should match to your component assign.
@@ -16,6 +16,8 @@ defmodule PhxLiveStorybook.Attr do
     * `:integer` - any integer
     * `:float` - any float
     * `:list` - a List of any arbitrary types
+    * `:block` - a block (you can only declare one)
+    * `:slots` - a list of slots
     * Any struct module
   - `required`: `true` if the attribute is mandatory.
   - `doc`: a text documentation for this attribute.