+++ test/fixtures/storybook_content/tree/a_folder/ab_component.exs
@@ -3,8 +3,22 @@ defmodule TreeStorybook.AFolder.AbComponent do
   def component, do: BComponent
   def description, do: "Ab component description"
 
+  def attributes do
+    [
+      %Attr{id: :label, type: :string, required: true},
+      %Attr{id: :block, type: :block}
+    ]
+  end
+
   def stories do
     [
+      %Story{
+        id: :default,
+        attributes: %{label: "hello"},
+        block: """
+        <span>inner block</span>
+        """
+      },
       %StoryGroup{
         id: :group,
         stories: [